# openapi.yaml
openapi: 3.1.0
info:
  title: CEOS ARD Editor API
  version: 0.1.0
  description: |
    RESTful API for the CEOS ARD Editor application that provides workspace management,
    file editing, building block management, and GitHub integration capabilities.
  contact:
    name: CEOS ARD Team
    url: https://github.com/ceos-org/ceos-ard-editor
  license:
    name: Apache-2.0
    url: https://github.com/ceos-org/ceos-ard-editor/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://ceos-ard-editor.example.com/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    get:
      summary: Initiate GitHub login
      description: Redirects to GitHub for authentication
      operationId: initiateLogin
      tags:
        - Authentication
      security: []
      responses:
        '302':
          description: Redirect to GitHub authentication
          headers:
            Location:
              description: GitHub authentication URL
              schema:
                type: string
                format: uri
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/logout:
    get:
      summary: Logout user
      description: Invalidates current session and tokens
      operationId: logout
      tags:
        - Authentication
      responses:
        '302':
          description: Successfully logged out
          headers:
            Location:
              description: Redirect URL after logout
              schema:
                type: string
                format: uri
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/user:
    get:
      summary: Get current user information
      description: Returns current authenticated user's information
      operationId: getCurrentUser
      tags:
        - Authentication
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
        - code
    User:
      type: object
      required:
        - id
        - username
        - displayName
        - email
        - avatar
      properties:
        id:
          type: string
          description: User ID
        username:
          type: string
          description: GitHub username
        displayName:
          type: string
          description: User's display name
        email:
          type: string
          format: email
          description: User's email address
        avatar:
          type: string
          format: uri
          description: URL to user's avatar image
...